Rename `Civi\Core\Theme` to `Civi\Core\Themes`
The class manages a list of themes -- not just a single theme.
