+++ CRM/Utils/Hook.php
@@ -532,13 +532,13 @@ abstract class CRM_Utils_Hook {
    *  - url_callback: mixed (optional)
    *         A function ($themes, $themeKey, $cssExt, $cssFile) which returns the URL(s) for a CSS resource.
    *         Returns either an array of URLs or PASSTHRU.
-   *         Ex: \Civi\Core\Theme\Resolvers::simple (default)
-   *         Ex: \Civi\Core\Theme\Resolvers::none
+   *         Ex: \Civi\Core\Themes\Resolvers::simple (default)
+   *         Ex: \Civi\Core\Themes\Resolvers::none
    *  - prefix: string (optional)
    *         A prefix within the extension folder to prepend to the file name.
    *  - search_order: array (optional)
    *         A list of themes to search.
-   *         Generally, the last theme should be "*fallback*" (Civi\Core\Theme::FALLBACK).
+   *         Generally, the last theme should be "*fallback*" (Civi\Core\Themes::FALLBACK).
    *  - excludes: array (optional)
    *         A list of files (eg "civicrm:css/bootstrap.css" or "$ext:$file") which should never
    *         be returned (they are excluded from display).
